Fix broken jql query
