+++ src/main/java/de/uhd/ifi/se/decision/management/bitbucket/oauth/JiraClient.java
@@ -1,5 +1,8 @@
 package de.uhd.ifi.se.decision.management.bitbucket.oauth;
 
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.Locale;
 import java.util.Set;
@@ -109,12 +112,27 @@ public interface JiraClient {
 		return !projectKey.isEmpty() && projectKeys.contains(projectKey);
 	}
 
-	public static String getJiraCallQuery(Set<String> messageList) {
+	public static String getJiraCallQuery(Set<String> jiraIssueKeys) {
 		String query = "?jql=key in (";
-		for (String message : messageList) {
-			query = query + message + ",";
+		Iterator<String> iterator = jiraIssueKeys.iterator();
+		while (iterator.hasNext()) {
+			String key = iterator.next();
+			query += key;
+			if (iterator.hasNext()) {
+				query += ",";
+			}
 		}
 		query += ")";
-		return query;
+		return encodeUserInputQuery(query);
+	}
+
+	private static String encodeUserInputQuery(String query) {
+		String encodedUrl = "";
+		try {
+			encodedUrl = URLEncoder.encode(query, "UTF-8");
+		} catch (UnsupportedEncodingException e) {
+			e.printStackTrace();
+		}
+		return encodedUrl;
 	}
 }
